mwaa-local-runner/workflows/requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.4.3/constraints-3.10.txt"


apache-airflow-providers-mysql==3.2.1
kubernetes==23.6.0
apache-airflow-providers-cncf-kubernetes==4.4.0
#cython==0.29.21
#pyarrow==2.0.0
#awswrangler==2.4.0
